requirements.txt
#devops
black==22.3.0
click==8.1.3 
pytest==7.1.3
pytest-cov==4.0.0
#pylint==2.15.3
#rust based linter
ruff==0.0.284
boto3==1.24.87
#web
fastapi==0.85.0
uvicorn==0.18.3
pandas==2.2.2
matplotlib==3.9.2
tabulate==0.9.0
nbval==0.11.0
polars-lts-cpu
pyarrow


